Temporarily fixed error preventing network from being deleted
The problem is that after network is deleted in delete_network_precommit,
OpenStack Ocata executes update_network_precommit which looks for deleted
network in Contrail. Contrail responds that it hasn't found this network
and the whole process is stopped.

It could be fixed by catching NetworkNotFound exception in
update_network_precommit, but in function which throws error a typo was
made (NetwrokNotFound), so NeutronException: An unknown exception is thrown
instead. We added temporary hack which detects this situation.